requirement.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
langchain
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
langchain_experimental

streamlit_chat